partner_ssn AFC-1562: broke social security number code out of partner_view_360
